Reviewer 2
Programmer,
I was not able to run the program properly. I kept getting error messages and when opening the app it would prompt me that the error comes from the very last line of code. It claims that the error is at the very end of the last line, however I couldn't identify the error or get it to load properly.  I tried to play around with a few things, but still couldn't get it to run on my emulator.  Later I received an error that said "[ERROR] Application Installer abnormal process termination. Process exit value was 8." I couldn't find a way past this either.  The code, however, seems to be correct.  It is entered in exactly how its written in the book, including the camera function.  I don't know why it wouldn't run properly, but the code seems to be correct.